+++ packages/database/src/schemas/index.ts
@@ -8,7 +8,6 @@ export * from './aiInfra';
 export * from './apiKey';
 export * from './asyncTask';
 export * from './betterAuth';
-export * from './brief';
 export * from './chatGroup';
 export * from './file';
 export * from './generation';
